Add a broken example of one attempt to define a generalized 'using' function.
This attempt almost works, but it will get a type error where close() is
called. The problem is that, as written, T can be any type at all. To
indicate that it only really supports types with close() methods,
the <: notation can be used to give an upper bound to T. In this case,
the desired upper bound is

{ def close(): Unit }.

Let's see the next commit for a complete working definition.