Add an example US CurrencyZone.
Here, US is an object that extends CurrencyZone. It defines a class
Dollar, which is a subclass of AbstractCurrency. So the type of money in
this zone is US.Dollar. The US object also fixes the type Currency to be
an alias for Dollar, and it gives an implementation of the make factory
method to return a dollar amount.

This is a workable design. There are only a few refinements to be added.