Add the notion of currency units and subunits.
So far, every currency was measured in a single unit: dollars, euros, or
yen. However, most currencies have subunits: for instance, in the US,
it's dollars and cents.

The most straightforward way to model cents is to have the amount field
in US.Currency represent cents instead of dollars. To convert back to
dollars, it’s useful to introduce a field CurrencyUnit into class
CurrencyZone, which contains the amount of one standard unit in that
currency.