Add an (arbitrary) example of an implicit conversion from Double to Int
Converting Doubles to Ints might raise some eyebrows, because it's a
dubious idea to have something that causes a loss in precision happen
invisibly. So this is not really a recommended conversion.

It makes much more sense to go the other way, from some more constrained
type to a more general one. For instance, an Int can be converted
without
loss of precision to a Double, so an implicit conversion from Int to
Double
makes sense.

In fact, that's exactly what happens. The scala.Predef object, which is
implicitly imported into every Scala program, defines implicit
conversions
that convert narrower numeric types to wider ones.

For instance, you will find in Predef the following conversion:

  implicit def int2double(x: Int): Double = x.toDouble

That's why in Scala Int values can be stored in variables of type
Double. There’s no special rule in the type system for this; it's just
an implicit conversion that gets applied.