Override head and tail of case class :: with constructor vals
This works because every case class parameter is implicitly also a field
of the class (it's like the parameter declaration was prefixed with
val).

Recall that Scala allows you to implement an abstract parameterless
method such as head or tail with a field.

So the code here directly uses the parameters head and tail as
implementations of the abstract methods head and tail that were
inherited from class List.