Add a second version of the RNA class, RNA2
This one lets you treat it more like a sequence of Base objects, so
functions like take and drop work with it.

Also worth noting is that since the stairway book was written, some
changes to the language have forced me to diverge from the source in the
book - I specifically needed to mix in mutable.IndexedSeqLike instead of
just regular old IndexedSeqLike. I had to override the update method as
well, which I did, I believe, the best I could, by extracting it from
the code in fromSeq and having it call RNA2.updated instead.