Add commentary explaining a potential discrepancy
The discrepancy is between the stated return type of Email.unapply,
Option[(String, String)] and the apparent violation of that return type
with the expression Some(user, domain).

It turns out that Scala will desugar an extra pair of parentheses when
you pass more than value to a function expecting a single argument
tuple.