Add an example of an extractor that supports vararg pattern matching
The question remains how an extractor can support vararg matching as
shown in the previous example, where patterns can have a varying number
of sub-patterns. The unapply methods encountered so far are not
sufficient, because they each return a fixed number of sub-elements in
the success case.

To handle this case, Scala lets you define a different
extraction method specifically for vararg matching. This method is
called unapplySeq.

The result type of an unapplySeq must conform to Option[Seq[T]].